requirements.txt
wheel==0.40.0
pydantic==1.10.7
fastapi==0.95.0
requests==2.28.2
requests-cache==1.0.1
uvicorn==0.21.1
gunicorn==20.1.0
setproctitle==1.3.2
itsdangerous==2.1.2
python-dotenv==1.0.0
